requirements.txt
# Core RAG + Embeddings
llama-index-core==0.14.2
llama-index-vector-stores-faiss==0.14.2
llama-index-llms-openai==0.14.2
llama-index-embeddings-huggingface==0.14.2

# Reranker
sentence-transformers==3.0.1

# Vector search backend
faiss-cpu==1.7.4

# UI
streamlit==1.37.0

# Utilities
python-dotenv==1.0.1
requests==2.32.3
